udpdate

+++ src/main/java/edu/indiana/soic/spidal/binary/tools/HistogramGenerator.java
@@ -88,7 +88,7 @@ public class HistogramGenerator {
                     }
                 }
             }else{
-                double pointsInRange = numPoints*(end-start+1);
+                double pointsInRange = (double)numPoints*(end-start+1);
                 double prob = numSamples/pointsInRange*100;
                 System.out.printf("Prob :" + prob ) ;
                 int count = 0;
